pyproject.toml
[project]
name = "clipboard-manager"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "keyboard>=0.13",
    "pillow>=10.0",
    "pyqt6>=6.7",
    "pytesseract>=0.3",
]

[project.scripts]
clipboard-manager = "clipboard_manager.main:main"

[tool.ruff]
target-version = "py312"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]

[tool.mypy]
strict = true
python_version = "3.12"
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short -q"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.20.1",
    "pytest>=9.0.3",
    "pytest-cov>=5.0",
    "ruff>=0.15.11",
]

[[tool.mypy.overrides]]
module = ["keyboard", "pytesseract"]
ignore_missing_imports = true
